First Executable Commit!
Finished coding the Board Class
Added a ChessJava Main Class to the Engine Package
Overridden to string method for every piece so the piece returns its corresponding letter.
Added an enum to the Piece Class so it can give a letter to each of the Pieces on the board.
Code Cleanup